+++ packages/_private/e2e-demo-test/src/AllComponents/SynSelect.spec.ts
@@ -166,7 +166,11 @@ test.describe('<SynSelect />', () => {
     }); // regression#885
 
     test.describe(`Regression#1036: ${name}`, () => {
-      test('should change value correct for subsequently changed delimiter', async ({ page }) => {
+      test('should change value correct for subsequently changed delimiter', async ({ browserName, page }) => {
+        // Unfortunately this test is flaky in firefox CI runs
+        // Works fine locally, but to keep the CI healthy we skip it there
+        test.skip(browserName === 'firefox', 'Flaky in firefox');
+
         const AllComponents = new AllComponentsPage(page, port);
         await AllComponents.loadInitialPage();
         await AllComponents.activateItem('selectLink');
@@ -180,9 +184,6 @@ test.describe('<SynSelect />', () => {
         const option2 = options.nth(1);
 
         await select.click();
-
-        // Wait for the first option to be visible
-        // Firefox is flaky without this statement
         await expect(select).toHaveAttribute('open');
         await expect(option1).toBeVisible();