+++ packages/components/src/components/side-nav/side-nav.styles.ts
@@ -32,9 +32,6 @@ export default css`
     position: absolute;
   }
 
-
-
-
   /****
    *
    * variant="default"
@@ -50,9 +47,6 @@ export default css`
   .side-nav--fix .side-nav__drawer::part(overlay){
     display: block;
   }
-
-
-
 
   /****
    *
@@ -88,9 +82,6 @@ export default css`
     overflow: hidden;
   }
 
-
-
-
   /****
    *
    * variant="sticky"
@@ -113,7 +104,6 @@ export default css`
   :not(.side-nav--open).side-nav--sticky .side-nav__drawer {
     --size: var(--side-nav-rail-width);
   }
-
 
   /*
    * For touch devices the content should not shrink in a flex container
@@ -144,7 +134,6 @@ export default css`
    ****/
   .side-nav__footer-divider{
     --spacing: 0;
-    --color: var(--syn-color-neutral-300);
   }
 
   :not(.side-nav--open).side-nav--rail ::slotted(syn-nav-item){